delete file generated by slice2cpp

+++ ConfigHelper.h
@@ -301,7 +301,7 @@ private:
         {
             XMLUtil::XmlNodes nodes = XMLUtil::locate(node, it_detail->path);
             if(nodes.size() != 1)
-                throwf<CfgException>(EXFMT(CfgException, "Holder::__readThis() bad xml definition, found %u nodes of path [%s]. root=%s"), nodes.size(), it_detail->path.c_str(), __m_rootPath.c_str());
+                throwf<CfgException>(EXFMT(CfgException, "Holder::__readThis() bad xml definition, found %u nodes of path [%s]. root=%s"), (unsigned int)nodes.size(), it_detail->path.c_str(), __m_rootPath.c_str());
 
             // read the config of this attribute
             XMLUtil::XmlNode target = nodes[0];
@@ -324,7 +324,7 @@ private:
                     else
                     {
                         throwf<CfgException>(EXFMT(CfgException, "Holder::__readThis() insufficient buffer. path [%s], attribute[%s], attribute size [%u], buffer size[%u]. root=%s")
-                            , it_detail->path.c_str(), it_detail->name.c_str(), str.size(), CFG_BUFSIZE, __m_rootPath.c_str());
+                            , it_detail->path.c_str(), it_detail->name.c_str(), (unsigned int)str.size(), (unsigned int)CFG_BUFSIZE, __m_rootPath.c_str());
                     }
                 }
                 value = buf;
@@ -353,7 +353,7 @@ private:
             if(target.size() < it_other->nodeCount.first || it_other->nodeCount.second < target.size())
             {
                 throwf<CfgException>(EXFMT(CfgException, "Holder::__readOthers() bad xml definition, found %u nodes of path [%s], violate the range[%u, %u]. root=%s"),
-                    target.size(), it_other->path.c_str(), it_other->nodeCount.first, it_other->nodeCount.second, __m_rootPath.c_str());
+                    (unsigned int)target.size(), it_other->path.c_str(), (unsigned int)it_other->nodeCount.first, (unsigned int)it_other->nodeCount.second, __m_rootPath.c_str());
             }
             for(XMLUtil::XmlNodes::iterator it_node = target.begin(); it_node != target.end(); ++it_node)
             {
@@ -421,7 +421,7 @@ private:
                 if(strlen(value) < _length)
                     strncpy(&(obj.*(PMem_CharArray)_address), value, _length);
                 else
-                    throwf<CfgException>(EXFMT(CfgException, "Detail::set() value to long, path[%s], name[%s], length limit[%u]. root=%s"), path.c_str(), name.c_str(), _length, root.c_str());
+                    throwf<CfgException>(EXFMT(CfgException, "Detail::set() value to long, path[%s], name[%s], length limit[%u]. root=%s"), path.c_str(), name.c_str(), (unsigned int)_length, root.c_str());
                 break;
             default:
                 throwf<CfgException>(EXFMT(CfgException, "Detail::set() bad data type, type[%d]"), _type);
